i#5645: Make the last rep/repne prefix "win" (#5648)
x86 allows multiple prefixes to be present on an instruction. Although not
documented in the Intel SDM, empirical testing shows that the last rep/repne
prefix "wins", since it makes no sense for both prefixes to be present.
Since these prefixes are used at times to select between instructions
(e.g. anything with PREFIX_EXT in the decoder table) faithfully replicating
this behavior is important to ensure that byte sequences with unnecessary
leading prefixes are correctly decoded.

Fixes #5645
Fixes #5447