i#5498: Fix drutil_insert_get_mem_addr for w28 index
The fix in PR #5511 for drutil_insert_get_mem_addr() on a w28 index
register failed to use the application value in a temp register and
resulted in an incorrect recorded address.  The test added in that PR
failed to actually check the value and so failed to catch the error.

Here we augment the test to check the value by passing the recorded
address and the operand to a clean call.  This new test fails for the
w28 case.  With the included fix, it now passes.

Fixes #5498