i#i5656 raw2trace: Delay function markers
Adds function markers to delayed_branch vector in
raw2trace_thread_data_t so that the markers are not inserted inside a
basic block (e.g: between an instruction and its memref).

Tested: Run raw2trace on a trace that previously had "Function marker
should be after a branch" invariant error and with this change that
invariant error goes away.

Fixes #5656