i#3544 riscv64, part 2: add instr_info_t generator
Add a script for parsing Instruction Set Listing files. The goal is to
generate:

1. A list of supported instruction opcodes (opcode_api.h).
2. A full set of instruction creation macros (instr_create_api.h).
3. An instr_info_t table and a lookup trie for all known instructions.

The first two files are generated based on template files so that they
may include non-generated data as well. The template files will be added
in following commits.
The last file is generated from scratch.

Issue #3544

Signed-off-by: Stanislaw Kardach <kda@semihalf.com>