i#5658 rseq drreg: Put native sequence by barrier (#5664)
Fixes PR #5663 which fixed drreg restoration for drbbdup but broke it
for non-drbbdup.  bb->instr was NULL so the preinsert was an append,
and the label as the last instruction moved the drreg restore point to
after the native sequence!

Here, we explicitly append to avoid confusion, and we move the native
sequence to the point of the barrier instead of at the final app
instr, which is where the registers are actually native.

Tested on larger apps using rseq which reliably crashed without this
fix.  This fix also seems to fully fix drbbdup usage with drmemtrace
for #2039.

Issue: #5658, #2039