i#5520 memtrace encodings: Ensure raw dir is not needed
Fixes the incorrect attempt by analyzer_interface to synthesize a
modules.log directory when running opcode_mix, which failes when there
is no raw directory.

Modifies the opcode_mix offline test to remove the raw directory as a
sanity check that the embedded instruction encodings are indeed all we
need.

Issue: #5520