i#3044: Add a type to represent aarch64 vectors
Previously in the IR we have represented vectors by a
plain H, S, D or Q register combined with a faux-imm
added after the last vector to give a hint to its size.

This patch uses the .size field of the opnd struct to
include the element size, similar to how partial registers
are used for x86. Element Vector registers are differentiated
from this by setting the DR_OPND_IS_VECTOR (0x40) bit on the
.flag field. This bit is also set by the DR_OPND_IS_EXTEND
flag for imms and was chosen as to not extend the size of the
flag field while remaining unambiguous.

The new Element Vector operand is printed like z0.b via a check
in disassemble_shared.c to set a suffix. Also added are the
utility functions opnd_is_element_vector_reg,
opnd_create_reg_element_vector and opnd_get_vector_element_size/

issue: #3044

Change-Id: I0cdb78cc1a13c3db7b6c742fb1d5d5c0e54216ff