i#2039 trace trim, part 2: Add open_file with tid
Adds drmemtrace_replace_file_ex_ops() with an expanded file opening
function which takes in the thread id and window id, to better support
the external file opener with delayed opens due to nop mode and
windows.

Adds a test of the function to burst_replaceall.  Adds additional
threads to the test and checks that the tid for each thread was seen.

Issue: #2039