i#2039 trace trim, part 2: Add open_file with tid (#5679)
Adds drmemtrace_replace_file_ex_ops() with an expanded file opening
function which takes in the thread id and window id, to better support
the external file opener with delayed opens due to nop mode and
windows.  The new function takes in a struct to make it easier to
extend further in the future.

Adds a test of the function to burst_replaceall.  Adds additional
threads to the test and checks that the tid for each thread was seen,
but disables that part on Windows due to #2040.

Issue: #2039