i#5520 memtrace encodings: Add online encodings (#5682)
Adds a new option -instr_encodings which enables instruction encoding
records for online traces.  This is under this off-by-default option
as it adds significant overhead to online tracing and is only needed
for some tools.

Currently we emit the encoding again for every dynamic instance
of an instruction.  Future work involves recording which we've emitted
to avoid duplicate instances (the reader caches prior encodings) but
this requires careful consideration of global locks and per-thread
invalidation on code changes which is out of the scope of this
initial implementation.

Updates the launcher and docs.

Adds an online opcode_mix test and adds encodings to a basic_counts
online test.

Manually tested the view tool with and without -instr_encodings:

  $ bin64/drrun -t drcachesim -instr_encodings -simulator_type view -- suite/tests/bin/allasm_x86_64
      ...
      166: T1718245 ifetch       4 byte(s) @ 0x0000000000401028 48 83 eb 01          sub    $0x0000000000000001 %rbx -> %rbx
      167: T1718245 ifetch       4 byte(s) @ 0x000000000040102c 48 83 fb 00          cmp    %rbx $0x0000000000000000
      168: T1718245 ifetch       2 byte(s) @ 0x0000000000401030 75 d9                jnz    $0x000000000040100b
      ...

  $ bin64/drrun -t drcachesim -simulator_type view -- suite/tests/bin/allasm_x86_64
      ...
      126: T1933188 ifetch       4 byte(s) @ 0x0000000000401028 non-branch
      127: T1933188 ifetch       4 byte(s) @ 0x000000000040102c non-branch
      128: T1933188 ifetch       2 byte(s) @ 0x0000000000401030 conditional jump
      ...

Fixes #5520