i#5686: Add restore state event to drbbdup
Previously drbbdup failed to restore the state for its own register
spills.  We add a restore state event here, for the case where we
have an ilist.

Tested on the drcacheoff.burst_threads test.  With a higher loop count
for the other threads, detach hits each instruction of the drbbdup
dispatch code with repeated runs and in manual testing I verified the
details of each.  It is difficult to construct a deterministic test of
asynch translation at these constrained points.

Issue: #5686