i#5434: Remove reliance on SIGSTKSZ and related defines
SIGSTKSZ and MINSIGSTKSZ are moving away from being constants.
SIGSTKSZ is now a call to sysconf().

This was causing a build error building our tests, which is fixed
using the suggestion from Stanislaw Kardach.

For the core, we eliminate the constants to avoid invoking libc.  We
look for AT_MINSIGSTKSIZE in our existing auxv search code and if not
found we fall back to a hardcoded value.

Co-authored-by: Stanislaw Kardach <kda@semihalf.com>
Fixes #5434