Convert reader_t and file_reader_t into templates.
This allows us to remove the second template parameter to analyzer_t, which
can now be expressed in terms of the first one.

However, we now have to duplicate some code in each implementation of file_reader_t
so that we support a version returning memref_t and one returning trace_entry_t.
This is because in C++ we cannot have partial specialization of member functions of
a template class.

The other option is to create a separate sub-class for each implementation of
file_reader_t.