i#5538 memtrace seek, part 7: Add instr count to tools (#5721)
Adds a new memtrace_stream_t interface class which provides the record
and instruction count to drmemtrace analysis tools.  A pointer to this
interface is passed to new extended-argument versions of
analysis_tool_t's initialize() (initialize_stream()) and 
parallel_shard_init() (parallel_shard_init_stream()) functions,
which are now what is called by the analyzer.  The base class
implementation of these new functions simply calls the old versions,
which are now deprecated but will continue to work.

This new interface is not just for convenience: the tool itself cannot
accurately count when the reader skips over records, as will happen
with seeking.  The counting must be done in the reader.  (If the tool
indeed wants to count only records/instrs that it actually sees, it
can continue using its own counters.)

We had considered other avenues for analysis_tool_t to obtain things like
the record and instruction ordinals within the stream, in the presence of
skipping: we could add fields to memref but we'd either have to append
and have them at different offsets for each type or we'd have to break
compatbility to prepend every time we added more; or we could add parameters
to process_memref().  Passing an interface to the init routines seems
the simplest and most flexible.

Updates the view tool to use the new interface to obtain the record
ordinal, replacing its own counter.  The tool is expanded to print a
new column with the instruction ordinal.  The view tool test is
updated along with example output in the docs.

Issue: #5538