i#5538 memtrace seek, part 9: Add accurate record count
Adds a reader subclass to raw2trace for computing the memref_t record
count for each chunk.  A new marker is inserted in the chunk header
which the zipfile skip code uses to obtain the correct ref count when
skipping over chunks.

Adds a count suppression feature where the view tool prints 0 for the
record count for the synthetic timestamp+cpu added after a seek.

Updates the seek test with a new trace and new expected output.

Issue: #5538