i#5383 mac a64, part 3: Add drsyms and test configurations (#5736)
Adds static libdwarf and libelftc libraries built for Mach-O AArch64.
These are not yet fully tested but this is a first pass to able
configuring and building drsyms and components that depend on it such
as drcachesim.  This enables using local M1 builds to replication GA
CI OSX build warnings.

Fixes build warnings and issues from things like the oldest OSX
supported, which we update to 11.7.
Eliminates syscall() from signal.dll.c as it is deprecated in 11.7.

Fixes configuration issues from x86-only tests.  Building the tests
does not yet succeed.

Issue: #5383