i#2039 trace trim, part 5: Drop early data (#5768)
Part 4 added a minimum timestamp, but it only works for data that is
not actually emitted until we are fully attached.  For data emitted
before that point, we'll use a too-early timestamp.  We solve that
here by dropping such data.

It is not easy to make an automated test unfortunately.

Issue: #2039