i#5763 ubuntu 22: fix gather into xmm test (#5764)
Fixes "gather result does not match" issue in scatter/gather tests. This was
because vpgatherqd and vgatherqps write only the first half of the xmm
reg when the index is an xmm reg, so the remaining part needs to be
explicitly initialized to zero.

With this fix, the scatter/gather tests pass on my local machine. On
Github Ubuntu-22 CI however, there is an unresolved SIGSEGV that
will be resolved separately. Added details to i#5763.

Issue: #5763