i#5675 record filter: Preserve encodings for filtered instrs
Adds logic to preserve encoding entries for instrs that were filtered
out. Such preserved encoding entries are delayed to next instrs for
their corresponding pc. This is required because TRACE_TYPE_ENCODING
entries must be immediately followed by an instr entry, since they
must be tied with a pc.

Extends the record_filter_unit_tests to verify encoding preservation.

We also want to add a unit test that uses a freshly generated trace
for filtering. But since the generated traces are .zip and we don't
have zip file writing support in record_filter, added a TODO for now.

Issue: #5675