i#5781: Fix RISCV64 compilation
Fixes compilation issues in the RISCV64 build, required due to recent
changes.

Due to the lack of RISC-V compilation in the CI the mainline development
did not notice breaking the compatibility. Therefore fix the compilation
issues:

- Exclude xflags handling in drbbdup for RISC-V64
- Add missing arch definition to drcachesim.

Note that this is only a stop-gap. The next action should be to add a
non-unit-test compilation to the CI. Non-unit-test because unit tests
don't yet compile on RISC-V.

Signed-off-by: Stanislaw Kardach <kda@semihalf.com>

Fixes: #5781