Add a new instr_t.offset field
Stops using instr_t.note to hold encoding offsets for pc-releative
operands.  Adds a new field instr_t.offset which is used for this
purpose.  This leaves note values in place across encodings, which is
needed for new clean call marking labels and also simplifies rseq
handling code.

This is a compatibility break and we should bump the version.

This should be committed separately; adding to this PR to ensure it
all works together.