i#5733, i#5734: Minor fixes in tests
Adds -max_trace_size to api.rseq to avoid generating too much trace data, which
may cause us to run out of disk space sometimes. The issue on varying run times
(and therefore trace data) is not reproducing now on the Jenkins machine, but
with this change, the trace size comes down to 512K from 130M+ which is
better.

Fixes an issue in the invariant checker that affects error reporting when there's
no shard stream available. This came up in the tool.drcacheoff.rseq test.

Issue: #5733, #5734