i#5733, i#5734: Minor fixes in tests (#5796)
Adds -max_trace_size to api.rseq to avoid generating too much trace data,
which may cause us to run out of disk space. The issue on varying run times
(and therefore trace data) is not reproducing now on the Jenkins machine,
but with this change, the trace size comes down to 512K from 130M+ which
is better.

Fixes an issue in the invariant checker that causes a SIGSEGV crash during
error reporting when there's no shard stream available. This came up in the
tool.drcacheoff.rseq test during the check_schedule_data checks that
happen at the very end in print_results of the invariant checker, where
we do not have the shard stream anymore. The invariant error itself is yet
to be debugged (i#5734). For the missing shard stream, we now use a
default stream with manually set ref ordinal.

Issue: #5733, #5734