i#5686 bbdup xl8: Modify xl8 params for other libs (#5831)
Moves the drbbdup restore state event to a very early priority and has
it modify the restore state parameters to focus on the containing copy
of the block where the restore resides.  It modifies the instruction
list and the cache start pc.

Adds a new test drx-scattergather-bbdup by adding drbbdup to the
existing drx-scattergather test.  This test fails to restore without
the fix described above, due to the scatter/gather expansion restore
state machine getting confused across multiple copies.

Does not augment drbbdup's restore to handle the no-ilist case: leaves
that for the general #3801 mechanism.

Fixes an uninitialized var compiler warning in the aarch64 codec change
PR #5821 #3044 b63dbdaa.

Issue: #5686