i#5538 memtrace seek, part 11: Fix omitted header problems (#5840)
Fixes a bug where reader_t's detection of duplicated timestamp,cpuid
headers at the start of a chunk assumed single-threaded mode.  We
switch to using a simple per-tid chunk footer trigger.

Adds a test to view_test via a new serial mock which takes in
trace_entry_t and allows testing of the interleaving code.  Tests both
proper chunk header elision as well as replicating the bug where
elision should not happen.

Fixes problems revealed in the drcacheoff.skip test by this change:
do not increment the ref count for the
hidden markers at the start of a chunk when skipping in a zipfile as
well as in raw2trace.

The test revealed a separate bug in the view tool where the version
and filetype ordinals, for delaying, were not updated on new threads.
That is fixed here as well as otherwise the new tests fail.

Issue: #5538