i#4489 drcachesim: support any cache associativity (#5867)
Remove restriction that cache associativity and size must be a
power-of-two. Updated requirement is that size / associativity must be a
power-of-two. Added tests for newly supported sizes, and for the updated
cache.init() parameter checkers.  Also renamed the blocks_per_set parameter
to the more accurate blocks_per_way.

Issue: #4489
Fixes: #4996