i#5784: Mitigate drwrap transparency violation
Adds a new drwrap API, drwrap_replace_if_retaddr_sentinel, that allows
mitigation of a transparency violation under the DRWRAP_REPLACE_RETADDR
drwrap strategy where the return address on the stack is replaced with
the address of the internal replace_retaddr_sentinel() routine. This
API modified the passed-in value to the actual return address of the
wrapped function if the passed-in value is replace_retaddr_sentinel()
itself.

Fixes the value of the marker written by the kernel xfer event in
drmemtrace by using the new drwrap_replace_if_retaddr_sentinel() API
on the mcontext PC before writing it out to the trace. This caused
many invariant errors of type 'Signal handler return point incorrect'
in traces collected on proprietary apps.

Verified on a large proprietary app that this error due to drwrap is
fixed now, whereas there were a few hundred instances before.

Fixes: #5784