i#5843 scheduler: Improve region skipping support
Fixes some fencepost errors in scheduler input region of interest handling.

Adds a test of regions of interest which actually contains timestamps,
which is what revealed the errors.

Refactors the scheduler unit tests to use trace_entry_t instead of
memref_t, which is required to properly test the scheduler's input
readers, as that is the record type they operate on.  This results in
no longer needing reader_t::use_prev() which is removed here.

Issue: #5843