Add input stream queries
Adds get_input_stream_count() and get_input_stream_name() to the
scheduler_t drmemtrace interface.

Adds a test of these to the scheduler unit tests which uses real files
and also serves as a test of only_threads for real files, whose code
paths are different enough it had a bug which we fix here as well.

Issue: #5843