Updates debug printing macro to include if().
The old DEBUG_VERSOSE(level) macro returns a condition that can be
evaluated at runtime for debug actions.  When not in DEBUG mode,
the condition always evaluates to false, resulting in unreachable
code the compiler will hopefully eliminate, but may cause a warning.

The new IF_DEBUG_VERBOSE(level, action) macro includes the
condition evaluation to perform <action> when the condition is
met in DEBUG mode.  When not in DEBUG mode, all code including
the condition is eliminated, leaving nothing to the compiler.