i#5724 chunk bounds: Fix incorrect instruction counts
Fixes a too-long chunk after a prior chunk splits up consecutive
delayed branches in the same buffer.

Avoids placing wasted encodings at the end of a chunk.

Adds a unit test, and reduces the chunk size for the larger test which
causes the original problem to reproduce much more reliably.

Still to fix is to re-emit encodings for the second consecutive branch
after the boundary.

Issue: #5724