i#5934 Add 'Repeated syscalls with the same PC' invariant check
Adds a dedicated invariant check for repeated syscalls that have the
same PC and reported as 'Non-explicit control flow has no marker'.

Tested with an app that has the 'Non-explicit control flow has no marker'
error for repeated syscalls and after this change, most of the errors
classified in that original error type are now reported as
'Repeated syscall instrs with the same PC'.

Fixes #5934