i#5724 chunk bounds: Insert encodings in new chunks
Adds logic to insert new encodings for instruction entries that were
generated before it was known they'd appear after a chunk boundary.
This has some inherent complexity due to the different modules owning
record generation and chunk creation, along with the complexity of
needing the decode PC for instruction records generated in the past.

Adds a unit test.

This finishes #5724, allowing us to remove it from the flaky list.

Fixes #5724