i#5934 dup syscall: Skip extraneous syscall in rn raw2trace.
Skips omitting to the final trace the 1-instr duplicate system call
block which has the same pc as the immediately previous system call
instr. This will serve as a workaround until we root-cause and fix
the underlying issue in the drmemtrace tracer.

Adds a unit test for this scenario.

Issue: #5934