i#5955: Update expected rseq offset for glibc 2.35
Glibc 2.35 made some changes to the __rseq_offset. We have logic that uses a
heuristic to search for the struct rseq offset. But we also assert on the
expected known rseq offset of glibc in the linux.rseq test; this is to
proactively detect glibc changes in rseq handling.

This fixes all rseq tests on glibc 2.35 except tool.drcacheoff.rseq which is
covered by #5734.

Fixes: #5955