i#5955: Update expected rseq offset for glibc 2.36 (#5956)
Adds struct rseq offset for glibc 2.36, while preserving the existing ones for
older glibc since our CI workflows are still on the older versions.

Glibc 2.36 made some changes to the __rseq_offset. We have logic that uses a
heuristic to search for the struct rseq offset. But we also assert on the
expected known glibc rseq offset in the linux.rseq test; this is to
proactively detect glibc changes in rseq handling.

This fixes all rseq tests on glibc 2.36 except tool.drcacheoff.rseq which is
covered by #5734.

Fixes: #5955