i#4041: Use rseq handler PC instead of commit PC (#5970)
The Linux kernel presents the rseq abort handler PC as the interrupted
(really, the resume point) for a signal generated while inside an rseq
region. Yet DR was modifying that to instead show clients the rseq
committing store PC. Similarly, DR presented that commit PC for its rseq
abort event if an rseq region aborted through some other non-signal
means like a thread migration. Since this is a little misleading and
doesn't match what the app sees, and results in the signal interrupted
PC not matching the later resume PC, we change that behavior here to
have both signal and rseq events use the handler PC. There is no way to
obtain the precise interruption PC from the kernel so we follow that
lead and do not try to present it from DR.

Furthemore, DR was not previously generating an rseq abort event if a
signal was the cause of the abort. We change that here, raising the
event prior to the signal event.

Updates raw2trace to use the handler PC that is now in the rseq
abort marker to detect an rseq abort, rather than looking for the
commit PC.

Adds a new test case to raw2trace_unit_tests and updates others.

Updates the api.rseq test.

Fixes #4041