i#5953 rseq: Fix rseq adjust filter and instru abort issues (#5971)
Fixes problems with the new rseq adjustment code with filtered traces:
+ Load and update the buffer pointer when emitting the rseq entry label
+ Do not try to adjust for i-filtered

Adds 2 new tests for filtering the linux.rseq app.

The new tests revealed a problem with the instr counts in the schedule
file: raw2trace was counting the PC-only entries as instructions; we fix that
here.

Reverts the rseq abort event and marker value back to the interrupted
PC for a signal in the instrumented execution.  Leaves all other cases
(all aborts in the native execution) as the handler.  This lets us
properly place the abort+signal for the instrumented interruption
case.
    
Adjusts the invariant checker abort check to use the new end PC for a
more accurate test to avoid false positives and negatives.

Fixes #5953, #4041