i#5984: Fix trace invariant false positive on back-to-back signals
Moves tracking for previous kernel_event marker values and previous pre-signal
instrs at the kernel_event and kernel_xfer markers. Previously we did some part
of this book-keeping at instrs after the kernel_xfer marker, which would be
erroneously skipped when there are back-to-back signals without any intervening
instrs.

Adds tracking for last seen instr at each signal depth. This is required to
correctly find the pre-signal instr for the second signal in a back-to-back
signal pair. Without this, the pre-signal instr for the second signal is
erroneously assumed to be whatever the last instr is, which may be some instr in
the first signal's body. This also obviates the need to track the boolean for
saw_kernel_xfer_after_prev_instr_, which was to detect nested signals Without
any intervening instr.

Adds a test case that verifies the scenario with back-to-back signals without
any intervening instr.

Fixes: #5984