i#5986: Remove timestamps too in rseq rollbacks
For rseq rollbacks in raw2trace, we can end up leaving a branch
followed by a timestamp after we remove other records.  This violates
our branch invariant.  We solve this by removing any timestamp
adjacent to the rollback region.

Adds a new unit test for this and adjusts an exiting one.

Further tested on a proprietary app where this showed up for real: it
is now gone.

Fixes #5986