i#5984: Fix trace invariant false positive on back-to-back signals (#5985)
Fixes some false positive invariant errors in the case where there are consecutive
signals without any intervening instr. Note that this is different than nested
signals without any intervening instr, which is already handled properly.

Moves tracking for previous kernel_event marker values and previous pre-signal
instrs to the kernel_event and kernel_xfer markers. Previously we did some part
of this book-keeping at instrs after the kernel_xfer marker, which would be
erroneously skipped when there are back-to-back signals without any intervening
instrs. Also created a separate struct and single stack for all such required state.

Improves tracking for last seen instr at each signal depth by making sure we
detect cases when there is no such instr. This is also required to correctly find the
pre-signal instr for the second signal in a consecutive signal pair without any
intervening instr. Without this, the pre-signal instr for the second signal is
erroneously assumed to be whatever the last instr is, which may be some instr
in the first signal's body. This also obviates the need to track the boolean for
saw_kernel_xfer_after_prev_instr_, which was to detect nested signals without
any intervening instr.

Handles the case where the trace starts inside a signal. The pre-signal instr check
and the kernel_event marker equality check both should be skipped in this case.

Handles the case where the trace starts inside a signal that returns and then has
a consecutive signal without any intervening instr. This requires some careful state
management to skip the pre-signal instr check for the second signal.

Adds detection of signals that caused an rseq abort, and skips the pre-signal
instr check for them, since the rseq abort redirects control to the abort
handler directly. This part was required in this PR due to the removal of the existing
saw_kernel_xfer_after_prev_instr_ check; prior to this PR, the
saw_kernel_xfer_after_prev_instr_ check fortuitously skipped the pre-signal
instr check because of the two consecutive kernel_event markers (consecutive
kernel_event markers generally mean back-to-back signals, but in this case the
first kernel_event is just an artifact of our rseq abort handling). This showed up
on the drcacheoff.rseq test. Modified an existing invariant checker unit test to
verify this scenario also.

Adds test cases that verify the handling of signal invariants in relevant scenarios
seen in actual large-app traces. Also tested manually on an actual large-app trace.

Fixes: #5984