i#5998: Do not install DR signal handlers in standalone mode
Avoids installing DR's own signal handlers when in standalone mode, as
we do not want to mess up the app's own handlers.  Standalone mode
does not support safe_read or try_except.

Tested on a large app with DR in standalone mode and confirmed
sigaction_syscall was not hit under the debugger any more.

Fixes #5998