i#5843 scheduler: Add record + replay
Adds recording of the dynamic schedule and replay of that recording.

Each thread segment is stored in an in-memory structure with the data
written out at the end to a specified ostream.

Replay uses a new mapping mode MAP_AS_PREVIOUSLY.  It reads from an
istream into memory, under the assumption that the schedule is small
enough to fit in memory.  The file format is an archive, however, to
future-proof it in case we need to read the cores incrementally in
parallel.  The record format has different needs from the existing
trace-recording schedule and so is a new structure.

Adds archive_istream.h and extends zipfile_istream.h to support
reading arbitrary zip files and advancing to specific subfiles.

The schedule data includes timestamps for more accurate re-creation of
the original inter-outupt-stream context switch timing, with an output
stream waiting if its timestamp gets ahead of other outputs.

Simultaneous reading of one input by multiple outputs is not
supported.  If one output stream gets ahead of another on the same
input stream, it has to wait for the other to catch up (which it would
do anyway based on the timestamp).

Waiting for another output currently has no kernel wait state support:
outputs are forced to spin and poll.

Regions of interest are handled with special skip entries in the
recorded schedule.  The synthetic records added during regular
scheduling are duplicating on replay.

Adds multiple unit tests.

A proprietary test setup with a real application trace was run under
ThreadSanitizer 500x to check for races.

Issue: #5843