i#5843 scheduler: Add replay of the as-traced schedule
Adds replay of the as-traced schedule using the cpu_schedule.zip file
written by raw2trace.  That file is converted into the record+replay
format for dynamic schedules and the existing replay code leveraged.

This implements as-traced cpu-oriented parallel iteration in the
scheduler which is part of #5694.  Adding analysis tool support for
this will be done separately.

Updates all replays (whether as-traced or as-previously) to consider
DEPENDENCY_TIMESTAMPS to indicate whether to have one output wait if
it gets ahead of another's timestamp.

Adds a missing check for output count mismatches on as-previously
replay.

Issue: #5843, #5694